Added a mechanism to detect when ralph enters/exits a room. This was done with quite a few changes:
Now, agent takes in an optional name argument, and a mandatory collisionHandler argument. This is so we can use the same collision handler for every agent, and also so we can add an event-generating pattern in the world. 

For the NPC, we fixed the issue where the key was tremendously huge, and when the character was insanely fast.

For the task timer, we used the globalClock object to detect the delta time in between ticks. This is supposed to be a better solution than keeping track of it ourselves.

The world now creates one collision handler for every agent. It also sets an event generating filter on the collision handler.